修复首页显示和文章详情中推荐的显示问题

+++ module_home/src/main/java/com/program/module_home/model/bean/ArticleRecommendBean.java
@@ -83,7 +83,7 @@ public class ArticleRecommendBean implements Serializable {
         @SerializedName("vip")
         private Boolean vip;
         @SerializedName("thumbUp")
-        private Double thumbUp;
+        private Integer thumbUp;
         @SerializedName("labels")
         private List<String> labels;
         @SerializedName("covers")
@@ -121,7 +121,7 @@ public class ArticleRecommendBean implements Serializable {
             return vip;
         }
 
-        public Double getThumbUp() {
+        public Integer getThumbUp() {
             return thumbUp;
         }
